requirements.txt
outline-vpn-api
python-decouple
python-dotenv
aiogram